nutriai/api/requirements.txt
# nutriai/api/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
sqlalchemy
google-generativeai
fpdf2